refactor: reorganize permission events to chain module
Move permission event schemas from types/events.ts to chain/permission0/permission0-events.ts
for better organization. Replace z.string().pipe(SS58_SCHEMA) with sb_address for consistency
with substrate parser patterns. Remove redundant parse wrapper functions in favor of direct
.safeParse() usage, simplifying the API and reducing unnecessary abstraction layers.
