+++ packages/torus-sdk-ts/src/types/index.ts
@@ -5,4 +5,4 @@ export * from "../chain/common/errors.js";
 export * from "./address.js";
 export * from "./namespace/index.js";
 
-export * from "./events.js";
+export * from "../chain/permission0/permission0-events.js";
